python-api/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
chromadb
sentence-transformers
python-dotenv
pypdf
cors
wikipedia
requests
